+++ Gymly/Workout/ShowSplitDayExerciseView.swift
@@ -17,7 +17,7 @@ struct ShowSplitDayExerciseView: View {
     @State var exercise: Exercise
     @State private var isOn = false
     @State var showSheet = false
-    @State var weight: Int = 0
+    @State var weight: Double = 0.0
     @State var reps: Int = 0
     @State var failure: Bool = false
     @State var warmUp: Bool = false
@@ -27,11 +27,11 @@ struct ShowSplitDayExerciseView: View {
     @State var setNumber: Int = 0
     @State var note: String = ""
     
-    var convertedWeight: Int {
+    var convertedWeight: Double {
         if config.weightUnit == "Kg" {
             return weight // Keep it as is
         } else {
-            return weight * Int(2.20462) // Convert Kg to Lbs
+            return weight * 2.20462 // Convert Kg to Lbs
         }
     }
     
@@ -163,7 +163,7 @@ struct ShowSplitDayExerciseView: View {
     func loadSetData(set: Exercise.Set, shouldOpenSheet: Bool = true) {
         if config.roundSetWeights {
             for index in exercise.sets.indices {
-                exercise.sets[index].weight = Int(exercise.sets[index].weight) // Ensure whole number rounding
+                exercise.sets[index].weight = exercise.sets[index].weight // Ensure whole number rounding
             }
             config.roundSetWeights = false
         }
